+++ src/sys.rs
@@ -220,8 +220,13 @@ pub struct RefpropLibrary {
     fn_critp: FnCritp,
     fn_trnprp: FnTrnprp,
     fn_setmix: FnSetmix,
+    fn_tdflsh: FnFlash,
+    fn_pdflsh: FnFlash,
     fn_thflsh: FnFlashKr,
     fn_tsflsh: FnFlashKr,
+    fn_dhflsh: FnFlash,
+    fn_dsflsh: FnFlash,
+    fn_hsflsh: FnFlash,
     fn_therm: FnTherm,
     fn_info: FnInfo,
 }
@@ -258,8 +263,13 @@ impl RefpropLibrary {
             fn_critp: Self::resolve(&lib, b"CRITPdll\0")?,
             fn_trnprp: Self::resolve(&lib, b"TRNPRPdll\0")?,
             fn_setmix: Self::resolve(&lib, b"SETMIXdll\0")?,
+            fn_tdflsh: Self::resolve(&lib, b"TDFLSHdll\0")?,
+            fn_pdflsh: Self::resolve(&lib, b"PDFLSHdll\0")?,
             fn_thflsh: Self::resolve(&lib, b"THFLSHdll\0")?,
             fn_tsflsh: Self::resolve(&lib, b"TSFLSHdll\0")?,
+            fn_dhflsh: Self::resolve(&lib, b"DHFLSHdll\0")?,
+            fn_dsflsh: Self::resolve(&lib, b"DSFLSHdll\0")?,
+            fn_hsflsh: Self::resolve(&lib, b"HSFLSHdll\0")?,
             fn_therm: Self::resolve(&lib, b"THERMdll\0")?,
             fn_info: Self::resolve(&lib, b"INFOdll\0")?,
             _lib: lib,
@@ -619,6 +629,98 @@ impl RefpropLibrary {
         }
     }
 
+    /// Temperature-density flash calculation.
+    pub unsafe fn TDFLSHdll(
+        &self,
+        t: *const c_double,
+        d: *const c_double,
+        z: *const c_double,
+        p: *mut c_double,
+        dl: *mut c_double,
+        dv: *mut c_double,
+        x: *mut c_double,
+        y: *mut c_double,
+        q: *mut c_double,
+        e: *mut c_double,
+        h: *mut c_double,
+        s: *mut c_double,
+        cv: *mut c_double,
+        cp: *mut c_double,
+        w: *mut c_double,
+        ierr: *mut c_int,
+        herr: *mut c_char,
+        herr_length: c_long,
+    ) {
+        unsafe {
+            (self.fn_tdflsh)(
+                t,
+                d,
+                z,
+                p,
+                dl,
+                dv,
+                x,
+                y,
+                q,
+                e,
+                h,
+                s,
+                cv,
+                cp,
+                w,
+                ierr,
+                herr,
+                herr_length,
+            );
+        }
+    }
+
+    /// Pressure-density flash calculation.
+    pub unsafe fn PDFLSHdll(
+        &self,
+        p: *const c_double,
+        d: *const c_double,
+        z: *const c_double,
+        t: *mut c_double,
+        dl: *mut c_double,
+        dv: *mut c_double,
+        x: *mut c_double,
+        y: *mut c_double,
+        q: *mut c_double,
+        e: *mut c_double,
+        h: *mut c_double,
+        s: *mut c_double,
+        cv: *mut c_double,
+        cp: *mut c_double,
+        w: *mut c_double,
+        ierr: *mut c_int,
+        herr: *mut c_char,
+        herr_length: c_long,
+    ) {
+        unsafe {
+            (self.fn_pdflsh)(
+                p,
+                d,
+                z,
+                t,
+                dl,
+                dv,
+                x,
+                y,
+                q,
+                e,
+                h,
+                s,
+                cv,
+                cp,
+                w,
+                ierr,
+                herr,
+                herr_length,
+            );
+        }
+    }
+
     /// Temperature-enthalpy flash calculation.
     pub unsafe fn THFLSHdll(
         &self,
@@ -705,6 +807,144 @@ impl RefpropLibrary {
                 q,
                 e,
                 h,
+                cv,
+                cp,
+                w,
+                ierr,
+                herr,
+                herr_length,
+            );
+        }
+    }
+
+    /// Density-enthalpy flash calculation.
+    pub unsafe fn DHFLSHdll(
+        &self,
+        d: *const c_double,
+        h: *const c_double,
+        z: *const c_double,
+        t: *mut c_double,
+        p: *mut c_double,
+        dl: *mut c_double,
+        dv: *mut c_double,
+        x: *mut c_double,
+        y: *mut c_double,
+        q: *mut c_double,
+        e: *mut c_double,
+        s: *mut c_double,
+        cv: *mut c_double,
+        cp: *mut c_double,
+        w: *mut c_double,
+        ierr: *mut c_int,
+        herr: *mut c_char,
+        herr_length: c_long,
+    ) {
+        unsafe {
+            (self.fn_dhflsh)(
+                d,
+                h,
+                z,
+                t,
+                p,
+                dl,
+                dv,
+                x,
+                y,
+                q,
+                e,
+                s,
+                cv,
+                cp,
+                w,
+                ierr,
+                herr,
+                herr_length,
+            );
+        }
+    }
+
+    /// Density-entropy flash calculation.
+    pub unsafe fn DSFLSHdll(
+        &self,
+        d: *const c_double,
+        s: *const c_double,
+        z: *const c_double,
+        t: *mut c_double,
+        p: *mut c_double,
+        dl: *mut c_double,
+        dv: *mut c_double,
+        x: *mut c_double,
+        y: *mut c_double,
+        q: *mut c_double,
+        e: *mut c_double,
+        h: *mut c_double,
+        cv: *mut c_double,
+        cp: *mut c_double,
+        w: *mut c_double,
+        ierr: *mut c_int,
+        herr: *mut c_char,
+        herr_length: c_long,
+    ) {
+        unsafe {
+            (self.fn_dsflsh)(
+                d,
+                s,
+                z,
+                t,
+                p,
+                dl,
+                dv,
+                x,
+                y,
+                q,
+                e,
+                h,
+                cv,
+                cp,
+                w,
+                ierr,
+                herr,
+                herr_length,
+            );
+        }
+    }
+
+    /// Enthalpy-entropy flash calculation.
+    pub unsafe fn HSFLSHdll(
+        &self,
+        h: *const c_double,
+        s: *const c_double,
+        z: *const c_double,
+        t: *mut c_double,
+        p: *mut c_double,
+        d: *mut c_double,
+        dl: *mut c_double,
+        dv: *mut c_double,
+        x: *mut c_double,
+        y: *mut c_double,
+        q: *mut c_double,
+        e: *mut c_double,
+        cv: *mut c_double,
+        cp: *mut c_double,
+        w: *mut c_double,
+        ierr: *mut c_int,
+        herr: *mut c_char,
+        herr_length: c_long,
+    ) {
+        unsafe {
+            (self.fn_hsflsh)(
+                h,
+                s,
+                z,
+                t,
+                p,
+                d,
+                dl,
+                dv,
+                x,
+                y,
+                q,
+                e,
                 cv,
                 cp,
                 w,